requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
google-generativeai==0.3.2
Pillow==10.2.0
python-dotenv==1.0.0
firebase-admin==6.4.0
